# docs/api/OPENAPI_SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: PwC Data Engineering Platform API
  description: |
    ## Enterprise-grade REST API for retail data analytics

    The PwC Data Engineering Platform API provides comprehensive access to retail analytics data, 
    advanced analytics capabilities, and business intelligence insights. Built with FastAPI, 
    this production-ready API includes:

    - **JWT Authentication** with role-based access control
    - **Rate Limiting** with intelligent burst protection  
    - **Circuit Breaker** patterns for resilience
    - **Multi-level Caching** for optimal performance
    - **Comprehensive Monitoring** with distributed tracing
    - **Data Quality** validation and scoring
    - **Advanced Analytics** with ML-powered insights

    ### Authentication
    This API uses JWT (JSON Web Tokens) for authentication. Include your token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ### Rate Limits
    - **Authentication endpoints**: 5 requests per minute
    - **Analytics endpoints**: 20 requests per minute (burst: 30)
    - **General endpoints**: 100 requests per minute (burst: 150)

    ### API Versions
    - **v1**: Stable CRUD operations and basic analytics
    - **v2**: Advanced analytics with ML capabilities
  version: 2.0.0
  termsOfService: https://pwc.com/terms
  contact:
    name: PwC Data Engineering Team
    email: data-engineering@pwc.com
    url: https://docs.pwc-data-platform.com
  license:
...
servers:
...
security:
  - BearerAuth: []
  - BasicAuth: []
...
paths:
...
  /api/v1/sales:
    get:
      tags: [Sales v1]
      summary: List sales transactions
      description: |
        Retrieve paginated sales transactions with comprehensive filtering options.
        Supports date range filtering, product/country filtering, and flexible sorting.
      parameters:
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
        - $ref: '#/components/parameters/Product'
        - $ref: '#/components/parameters/Country'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: List of sales transactions
          headers:
            X-Total-Count:
              description: Total number of records
              schema:
                type: integer
            X-Page-Count:
              description: Total number of pages
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSales'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /api/v1/sales/{transaction_id}:
    get:
      tags: [Sales v1]
      summary: Get sales transaction by ID
      description: Retrieve detailed information for a specific sales transaction
      parameters:
        - name: transaction_id
          in: path
          required: true
          description: Unique transaction identifier
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Sales transaction details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaleItem'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/sales/analytics:
    get:
      tags: [Sales v1]
      summary: Basic sales analytics
      description: Get aggregated sales metrics and key performance indicators
      parameters:
        - name: metrics
          in: query
          description: Metrics to calculate
          schema:
            type: array
            items:
              type: string
              enum: [revenue, transactions, customers, products]
            example: [revenue, transactions, customers]
        - name: dimensions
          in: query
          description: Dimensions to group by
          schema:
            type: array
            items:
              type: string
              enum: [country, product, time, customer_segment]
            example: [country, time]
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
        - name: granularity
          in: query
          description: Time granularity for time-based analysis
          schema:
            type: string
            enum: [hour, day, week, month, quarter, year]
            default: day
      responses:
        '200':
          description: Sales analytics data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesAnalytics'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Sales v2 Endpoints (Enhanced)
  /api/v2/sales:
    get:
      tags: [Sales v2]
      summary: Enhanced sales data with ML features
      description: |
        Advanced sales data endpoint with machine learning enrichment, 
        customer segmentation, and predictive features.
      parameters:
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
        - $ref: '#/components/parameters/Product'
        - $ref: '#/components/parameters/Country'
        - name: include_ml_features
          in: query
          description: Include machine learning features in response
          schema:
            type: boolean
            default: false
        - name: customer_segment
          in: query
          description: Filter by customer segment
          schema:
            type: string
            enum: [new_customers, loyal_customers, at_risk, champions]
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Enhanced sales data with ML features
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnhancedPaginatedSales'

  # Advanced Analytics v2 Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/v1/auth/login

    BasicAuth:
      type: http
      scheme: basic
      description: Basic HTTP authentication (fallback for development)
  parameters:
    DateFrom:
      name: date_from
      in: query
      description: Start date filter (ISO 8601 format)
      schema:
        type: string
        format: date
        example: "2024-01-01"
    DateTo:
      name: date_to
      in: query
      description: End date filter (ISO 8601 format)
      schema:
        type: string
        format: date
        example: "2024-01-31"
    Product:
      name: product
      in: query
      description: Product code filter
      schema:
        type: string
        example: "85123A"
    Country:
      name: country
      in: query
      description: Country name filter
      schema:
        type: string
        example: "United Kingdom"
    Page:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
    Size:
      name: size
      in: query
      description: Page size
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 20
        example: 100
    Sort:
      name: sort
      in: query
      description: Sort field and direction (field:asc|desc)
      schema:
        type: string
        default: "invoice_date:desc"
        example: "total_amount:desc"
  schemas:
...
    SaleItem:
      type: object
      required:
        - invoice_no
        - stock_code
        - quantity
        - invoice_date
        - unit_price
        - total
      properties:
        transaction_id:
          type: string
          format: uuid
          description: Unique transaction identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        invoice_no:
          type: string
          description: Invoice number
          example: "536365"
        stock_code:
          type: string
          description: Product stock code
          example: "85123A"
        description:
          type: string
          nullable: true
          description: Product description
          example: "WHITE HANGING HEART T-LIGHT HOLDER"
        quantity:
          type: integer
          description: Quantity purchased
          example: 6
        invoice_date:
          type: string
          format: date-time
          description: Invoice date and time
          example: "2024-01-01T08:26:00Z"
        unit_price:
          type: number
          format: float
          description: Price per unit
          example: 2.55
        customer_id:
          type: string
          nullable: true
          description: Customer identifier
          example: "17850"
        country:
          type: string
          nullable: true
          description: Customer country
          example: "United Kingdom"
        total:
          type: number
          format: float
          description: Total amount (quantity × unit_price)
          example: 15.30
        total_str:
          type: string
          description: Formatted total amount
          example: "15.30"
        data_quality_score:
          type: number
          format: float
          description: Data quality score (0.0-1.0)
          example: 0.95
        created_at:
          type: string
          format: date-time
          description: Record creation timestamp
          example: "2024-01-01T08:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Record last update timestamp
          example: "2024-01-01T08:30:00Z"
    PaginatedSales:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SaleItem'
        total:
          type: integer
          description: Total number of records
          example: 1000
        page:
          type: integer
          description: Current page number
          example: 1
        size:
          type: integer
          description: Page size
          example: 20
        pages:
          type: integer
          description: Total number of pages
          example: 50
    EnhancedPaginatedSales:
      allOf:
        - $ref: '#/components/schemas/PaginatedSales'
        - type: object
          properties:
            items:
              type: array
              items:
                allOf:
                  - $ref: '#/components/schemas/SaleItem'
                  - type: object
                    properties:
                      ml_features:
                        type: object
                        description: Machine learning enriched features
                        properties:
                          customer_segment:
                            type: string
                            enum: [new_customers, loyal_customers, at_risk, champions]
                            example: "loyal_customers"
                          predicted_clv:
                            type: number
                            format: float
                            description: Predicted customer lifetime value
                            example: 1250.75
                          churn_probability:
                            type: number
                            format: float
                            description: Probability of customer churn (0.0-1.0)
                            example: 0.15
                          product_affinity:
                            type: array
                            items:
                              type: object
                              properties:
                                product_code:
                                  type: string
                                  example: "85123B"
                                affinity_score:
                                  type: number
                                  format: float
                                  example: 0.85
            aggregations:
              type: object
              description: Aggregated statistics for current filter
              properties:
                total_revenue:
                  type: number
                  format: float
                  example: 125000.50
                unique_customers:
                  type: integer
                  example: 8945
                avg_order_value:
                  type: number
                  format: float
                  example: 27.63

    # Analytics Schemas
    SalesAnalytics:
      type: object
      properties:
        metrics:
          type: object
          properties:
            total_revenue:
              type: number
              format: float
              example: 1250000.50
            total_transactions:
              type: integer
              example: 45230
            unique_customers:
              type: integer
              example: 8945
            unique_products:
              type: integer
              example: 3500
            avg_order_value:
              type: number
              format: float
              example: 27.63
        dimensions:
          type: object
          properties:
            by_country:
              type: array
              items:
                type: object
                properties:
                  country:
                    type: string
                    example: "United Kingdom"
                  revenue:
                    type: number
                    format: float
                    example: 650000.25
                  transactions:
                    type: integer
                    example: 15000
            by_product:
              type: array
              items:
                type: object
                properties:
                  product_code:
                    type: string
                    example: "85123A"
                  description:
                    type: string
                    example: "WHITE HANGING HEART T-LIGHT HOLDER"
                  revenue:
                    type: number
                    format: float
                    example: 25000.75
                  quantity_sold:
                    type: integer
                    example: 1500
            by_time:
              type: array
              items:
                type: object
                properties:
                  date:
                    type: string
                    format: date
                    example: "2024-01-01"
                  revenue:
                    type: number
                    format: float
                    example: 15000.50
                  transactions:
                    type: integer
                    example: 450
        period:
          type: object
          properties:
            start_date:
              type: string
              format: date
              example: "2024-01-01"
            end_date:
              type: string
              format: date
              example: "2024-01-31"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error message
              example: "Request validation failed"
            details:
              type: object
              description: Additional error details
              properties:
                field:
                  type: string
                  example: "customer_id"
                message:
                  type: string
                  example: "Field is required"
            timestamp:
              type: string
              format: date-time
              example: "2024-01-15T10:30:00Z"
            request_id:
              type: string
              description: Unique request identifier for troubleshooting
              example: "req_abc123"
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Authentication required"
              timestamp: "2024-01-15T10:30:00Z"
              request_id: "req_abc123"
...
    RateLimitError:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Time when rate limit resets (Unix timestamp)
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retry
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "RATE_LIMIT_EXCEEDED"
              message: "Rate limit exceeded. Please retry after 60 seconds."
              timestamp: "2024-01-15T10:30:00Z"
              request_id: "req_abc123"
...